Module glitter::buffer

Exposes the OpenGL Buffer object, and related types.

Structs

Buffer
An OpenGL buffer object.

Enums

BufferBindingTarget
All of the possible OpenGL targets for binding a buffer object.

BufferDataUsage
Provides a hint to the OpenGL driver for how a buffer will be used.

Constants

ARRAY_BUFFER
The array buffer object binding.

DYNAMIC_DRAW
Indicates that a buffer will be set and modified repeatedly and drawn many times.

ELEMENT_ARRAY_BUFFER
The element array buffer object binding.

STATIC_DRAW
Indicates that a buffer will be set once and drawn many times

STREAM_DRAW
Indicates that a buffer will be set once and drawn only a few times.