Struct glitter::buffer::Buffer

pub struct Buffer {
    // some fields omitted
}

An OpenGL buffer object.

In OpenGL, a buffer is essentially untyped, and what data it can contain only depends on how it is used. A buffer can be used to store vertices, indices, or essentially any other data that the graphics device needs to process.

A buffer will automatically be deleted after going out of scope.

See also

gl.gen_buffer - Create a new, empty buffer.

VertexBuffer and IndexBuffer: Wrapper types over raw buffers that are more suitable for buffering vertex and index data, respectively.

gl.bind_array_buffer: and gl.bind_element_array_buffer: Bind a buffer to a target, returning a buffer binding type.