Module glitter::context

Home of ContextOf, which is the type that represents "the OpenGL state machine", and the type you use to make OpenGL calls.

Reexports

pub use self::buffer_context::*;

pub use self::framebuffer_context::*;

pub use self::program_context::*;

pub use self::renderbuffer_context::*;

pub use self::texture_context::*;

pub use self::texture_units::*;

Modules

buffer_context
Contains all of the OpenGL state types related to buffer objects.

ext
Contains all of the ContextOf extension traits that implement core OpenGL functionality.

framebuffer_context
Contains all of the OpenGL state types related to framebuffer objects.

program_context
Contains all of the OpenGL state types related to shader programs.

renderbuffer_context
Contains all of the OpenGL state types related to renderbuffers.

texture_context
Contains all of the OpenGL state types related to texture bindings.

texture_units
Contains all of the OpenGL state types related texture units.

Structs

ContextOf
The type that represents the whole "OpenGL state machine". This is the core of glitter's design, and what enables the notion of safety.

Traits

AContext
The 'core' OpenGL context trait. This trait provides access to any OpenGL functionality that don't deal with binding. This trait is implemented for ContextOf<_, _, _, _, _>, as well as for &mut ContextOf<_, _, _, _, _>.

BaseContext
A marker trait for types that represent an active OpenGL context.

ContextExt
An extension trait that contains some of the core OpenGL methods that maintain state, such as the current clear color or whether depth testing is enabled.

Type Definitions

Context
A "fresh" OpenGL context: one that essentially has no active bindings. See the ContextOf docs for more details.