Module glitter::context::buffer_context

Contains all of the OpenGL state types related to buffer objects.

Structs

ArrayBufferBinder
The OpenGL state representing the GL_ARRAY_BUFFER target.

ArrayBufferBinding
Represents a buffer that has been bound to the GL_ARRAY_BUFFER binding target.

BufferBinderOf
This type holds all of the OpenGL state releated buffer objects. See the ContextOf docs for more details.

ElementArrayBufferBinder
The OpenGL state representing the GL_ELEMENT_ARRAY_BUFFER target.

ElementArrayBufferBinding
Represents a buffer that has been bound to the GL_ELEMENT_ARRAY_BUFFER binding target.

Traits

ArrayBufferContext
An OpenGL context that has a free GL_ARRAY_BUFFER binding.

BufferBinding
A buffer that has been bound to an OpenGL buffer binding point.

BufferContext
An OpenGL context that has all free buffer bindings. This trait implies both ArrayBufferContext and ElementArrayBufferContext.

ContextBufferExt
An extension trait that includes buffer object-related OpenGL methods.

ElementArrayBufferContext
An OpenGL context that has a free GL_ELEMENT_ARRAY_BUFFER binding.

Type Definitions

BufferBinder
A part of the OpenGL context that has all free buffer bindings.