Module glitter::context::framebuffer_context

Contains all of the OpenGL state types related to framebuffer objects.

Structs

FramebufferBinder
The OpenGL state representing the GL_FRAMEBUFFER target.

FramebufferBinding
Represents a framebuffer that has been bound to the GL_FRAMEBUFFER binding target.

FramebufferBuilder
Provides a safe interface for building a framebuffer object that is checked to be framebuffer-complete. A FramebufferBuilder can be created using the gl.build_framebuffer method.

Enums

FramebufferAttachment
The various attachment points of a framebuffer object.

FramebufferTarget
All of the possible OpenGL targets for binding framebuffer objects.

Constants

COLOR_ATTACHMENT0
The color buffer attachment point.

DEPTH_ATTACHMENT
The depth buffer attachment point.

FRAMEBUFFER
The lone framebuffer target.

STENCIL_ATTACHMENT
The stencil buffer attachment point.

Traits

ContextFramebufferBuilderExt
The extension trait for contexts that adds the build_framebuffer method. This trait is only implemented for contexts that have a free framebuffer binding.

ContextFramebufferExt
An extension trait that includes framebuffer-related OpenGL methods.

FramebufferContext
An OpenGL context that has a free GL_FRAMEBUFFER binding.