Trait glitter::context::framebuffer_context::ContextFramebufferBuilderExt

pub trait ContextFramebufferBuilderExt: FramebufferContext + Sized {
    fn build_framebuffer<'a>(self) -> FramebufferBuilder<'a, Self> { ... }
}

The extension trait for contexts that adds the build_framebuffer method. This trait is only implemented for contexts that have a free framebuffer binding.

Provided Methods

fn build_framebuffer<'a>(self) -> FramebufferBuilder<'a, Self>

Create a new framebuffer builder, providing a safe interface for constructing a framebuffer object. See the FramebufferBuilder docs for more details.