Trait glitter::context::framebuffer_context::FramebufferContext

pub trait FramebufferContext: AContext {
    type Binder: BorrowMut<FramebufferBinder>;
    type Rest: AContext;
    fn split_framebuffer(self) -> (Self::Binder, Self::Rest);

    fn bind_framebuffer<'a>(
        self, 
        fbo: &'a mut Framebuffer
    ) -> (FramebufferBinding<'a>, Self::Rest)
    where
        Self: Sized,
    { ... }
}

An OpenGL context that has a free GL_FRAMEBUFFER binding.

Associated Types

type Binder: BorrowMut<FramebufferBinder>

The type of binder this context contains.

type Rest: AContext

The OpenGL context that will be returned after binding a framebuffer.

Required Methods

fn split_framebuffer(self) -> (Self::Binder, Self::Rest)

Split the context into a binder and the remaining context.

Provided Methods

fn bind_framebuffer<'a>(
self,
fbo: &'a mut Framebuffer
) -> (FramebufferBinding<'a>, Self::Rest) where
Self: Sized,

Bind a buffer to this context's framebuffer, returning a new context and a binding.

See also

glBindFramebuffer OpenGL docs