Module glitter::context::program_context

Contains all of the OpenGL state types related to shader programs.

Structs

ProgramBinder
The OpenGL state representing the active program target.

ProgramBinding
Represents a program that has been bound to the context.

ProgramBuilder
Provides a safe interface for creating program objects. A ProgramBuilder can be created using the gl.build_program method.

UnknownProgramAttrib
An error that represents a program attribute that could not be found.

UnknownProgramUniform
An error that represents a program uniform that could not be found.

Traits

ContextProgramBuilderExt
The extension trait that adds the build_program method.

ContextProgramExt
An extension trait that includes program-related OpenGL methods.

ProgramContext
An OpenGL context that has a free program binding.