Module glitter::context::renderbuffer_context

Contains all of the OpenGL state types related to renderbuffers.

Structs

RenderbufferBinder
The OpenGL state representing the GL_RENDERBUFFER target.

RenderbufferBinding
Represents a renderbuffer that has been bound to the GL_RENDERBUFFER binding target.

RenderbufferBuilder
Provides a safe wrapper for creating renderbuffer objects. A RenderbufferBuilder can be created using the gl.build_renderbuffer method.

Traits

ContextRenderbufferBuilderExt
The extension trait for contexts that adds the build_renderbuffer method. This trait is only implemented for contexts with a free renderbuffer binding.

ContextRenderbufferExt
An extension trait that includes renderbuffer-related OpenGL methods.

RenderbufferContext
An OpenGL context that has a free GL_RENDERBUFFER binding.