Struct glitter::context::renderbuffer_context::RenderbufferBuilder

pub struct RenderbufferBuilder<C> where C: RenderbufferContext {
    // some fields omitted
}

Provides a safe wrapper for creating renderbuffer objects. A RenderbufferBuilder can be created using the gl.build_renderbuffer method.

Methods

impl<C> RenderbufferBuilder<C> where C: RenderbufferContext

fn storage(self, format: RenderbufferFormat, width: u32, height: u32) -> Self

Set the storage parameters for the renderbuffer.

fn try_unwrap(self) -> Result<Renderbuffer, GLError>

Create and return a renderbuffer with the provided storage options, or return an error.

Failures

An error will be returned if no storage options were provided.

Panics

This function will panic if an OpenGL error is generated and debug assertions are enabled.

fn unwrap(self) -> Renderbuffer

Create a renderbuffer with the provided storage options, or panic.

Panics

This function will panic if no storage options were provided or if an OpenGL error was generated and debug assertions are enabled.