Trait glitter::context::renderbuffer_context::ContextRenderbufferExt

pub trait ContextRenderbufferExt: BaseContext {
    unsafe fn gen_renderbuffer(&self) -> Renderbuffer { ... }
    fn storage(&self, gl_rbo: &mut RenderbufferBinding, format: RenderbufferFormat, width: u32, height: u32) { ... }
}

An extension trait that includes renderbuffer-related OpenGL methods.

Provided Methods

unsafe fn gen_renderbuffer(&self) -> Renderbuffer

Create a new renderbuffer object with no storage allocated.

Safety

Many OpenGL function calls expect a renderbuffer to be attachment-complete. Be sure to properly set up the renderbuffer's storage before passing the renderbuffer to such functions.

fn storage(&self, gl_rbo: &mut RenderbufferBinding, format: RenderbufferFormat, width: u32, height: u32)

Initialize a renderbuffer object's storage.

gl_rbo: The binding of the renderbuffer to set up storage for.

format: The storage format to use for the renderbuffer.

width: The storage width of the renderbuffer, in pixels.

height: The storage height of the renderbuffer, in pixels.

See also

glRenderbufferStorage OpenGL docs