Module glitter::context::texture_context

Contains all of the OpenGL state types related to texture bindings.

Structs

Texture2dBinder
The OpenGL texture unit state that represents the GL_TEXTURE_2D target.

Texture2dBinding
Represents a texture that has been bound to the GL_TEXTURE_2D binding target of a texture unit.

Texture2dBuilder
Provide a safe interface for building a 2D texture object that is checked to be complete. A Texture2dBuilder can be created using the gl.build_texture_2d method.

TextureCubeMapBinder
The OpenGL texture unit state that represents the GL_TEXTURE_CUBE_MAP target.

TextureCubeMapBinding
Represents a texture that has been bound to the GL_TEXTURE_CUBE_MAP binding target of a texture unit.

Traits

ContextTextureBuilderExt
The extension trait for contexts that adds the build_texture_2d method.

ContextTextureExt
An extension trait that includes texture-related OpenGL methods.

TextureBinding
Represents a texture that has been bound to a texture unit.