Trait glitter::context::texture_context::ContextTextureBuilderExt

pub trait ContextTextureBuilderExt: TextureUnit0Context + Sized {
    fn build_texture_2d<'a>(self) -> Texture2dBuilder<'a, Self> { ... }
}

The extension trait for contexts that adds the build_texture_2d method.

Note

Currently, this trait is only implemented for contexts where the 0th texture unit is free.

Provided Methods

fn build_texture_2d<'a>(self) -> Texture2dBuilder<'a, Self>

Create a new 2D texture builder, providing a safe interface for constructing a 2D texture object. See the Texture2dBuilder docs for more details.