Module glitter::context::texture_units

Contains all of the OpenGL state types related texture units.

Structs

TextureSampler
A newtype wrapper representing a texture sampler, which can be used to set a uniform variable, using gl.set_uniform.

TextureUnit0
The 0th texture unit.

TextureUnit1
The 1st texture unit.

TextureUnit2
The 2nd texture unit.

TextureUnit3
The 3rd texture unit.

TextureUnit4
The 4th texture unit.

TextureUnit5
The 5th texture unit.

TextureUnit6
The 6th texture unit.

TextureUnit7
The 7th texture unit.

TextureUnitBindingOf
A texture unit that has been made active, and can have textures bound to it.

TextureUnitsOf
This type holds all of the OpenGL textrure units. Each type parameter is the current type of a texture unit. See the ContextOf docs for more details.

Traits

ATextureUnitBinding
A marker trait for types that represent an active texture unit binding.

TextureUnit
A trait that represents a 'texture unit', which is a piece of OpenGL state that contains its own independent texture bindings.

TextureUnit0Context
An OpenGL context with the 0th texture unit free.

TextureUnit1Context
An OpenGL context with the 1st texture unit free.

TextureUnit2Context
An OpenGL context with the 2nd texture unit free.

TextureUnit3Context
An OpenGL context with the 3rd texture unit free.

TextureUnit4Context
An OpenGL context with the 4th texture unit free.

TextureUnit5Context
An OpenGL context with the 5th texture unit free.

TextureUnit6Context
An OpenGL context with the 6th texture unit free.

TextureUnit7Context
An OpenGL context with the 7th texture unit free.

TextureUnitBinding2d
A texture unit binding that has a free GL_TEXTURE_2D binding.

TextureUnitBindingCubeMap
A texture unit binding that has a free GL_TEXTURE_CUBE_MAP binding.

Type Definitions

TextureUnitBinding
A fresh texture unit binding, that has all free texture bindings.

TextureUnits
A part of the OpenGL context that has all free texture units.