Trait glitter::context::texture_units::ATextureUnitBinding

pub unsafe trait ATextureUnitBinding { }

A marker trait for types that represent an active texture unit binding.

Safety

This type should only be implemented for types that can guarantee that an active texture unit will be bound for the lifetime of an instance of this type.