Trait glitter::context::texture_units::TextureUnit

pub trait TextureUnit {
    fn idx(&self) -> u32;

    fn active(&mut self) -> TextureUnitBinding { ... }
}

A trait that represents a 'texture unit', which is a piece of OpenGL state that contains its own independent texture bindings.

Required Methods

fn idx(&self) -> u32

Get the index of the texture unit.

Provided Methods

fn active(&mut self) -> TextureUnitBinding

Make the current texture unit active, returning a binding.