Trait glitter::context::texture_units::TextureUnit0Context

pub trait TextureUnit0Context: AContext {
    type Unit: BorrowMut<TextureUnit0>;
    type Rest: AContext;
    fn split_tex_unit_0(self) -> (Self::Unit, Self::Rest);

    fn active_texture_0(self) -> (TextureUnitBinding, Self::Rest) where Self: Sized { ... }
}

An OpenGL context with the 0th texture unit free.

Associated Types

type Unit: BorrowMut<TextureUnit0>

The type of unit this context contains.

type Rest: AContext

The OpenGL context that will be returned after making the texture unit active.

Required Methods

fn split_tex_unit_0(self) -> (Self::Unit, Self::Rest)

Split the 0th texture unit from the context, returning the unit and the remaining context.

Provided Methods

fn active_texture_0(self) -> (TextureUnitBinding, Self::Rest) where Self: Sized

Make the 0th texture unit active, returning a binding and the remaining context