Trait glitter::context::texture_units::TextureUnit1Context

pub trait TextureUnit1Context: AContext {
    type Unit: BorrowMut<TextureUnit1>;
    type Rest: AContext;
    fn split_tex_unit_1(self) -> (Self::Unit, Self::Rest);

    fn active_texture_1(self) -> (TextureUnitBinding, Self::Rest)
    where
        Self: Sized,
    { ... }
}

An OpenGL context with the 1st texture unit free.

Associated Types

type Unit: BorrowMut<TextureUnit1>

The type of unit this context contains.

type Rest: AContext

The OpenGL context that will be returned after making the texture unit active.

Required Methods

fn split_tex_unit_1(self) -> (Self::Unit, Self::Rest)

Split the 1st texture unit from the context, returning the unit and the remaining context.

Provided Methods

fn active_texture_1(self) -> (TextureUnitBinding, Self::Rest) where
Self: Sized,

Make the 1st texture unit active, returning a binding and the remaining context