Module glitter::image_data

Contains types related to 2D image data.

Structs

ImageFormat
A type that contains the texel type and format that make up a 2D image.

Pixel
A single OpenGL color value, with u8 components laid out as a C struct in RGBA order. This type is the simplest implementation of Image2d, which allows it to be uploaded to a texture.

Pixels
A (heap-allocated) 2D image composed of a list of pixels.

Enums

RenderbufferFormat
The various image formats of a renderbuffer.

TexelFormat
The different texel formats.

TexelType
The data types for a texel.

Constants

ALPHA
A texel contains only an alpha component.

DEPTH_COMPONENT16
The renderbuffer stores a 16-bit depth component.

RGB
A texel contains red, green, and blue components.

RGB565
The red, green, and blue channels are stored with 5 bits, 6 bits, and 5 bits, respectively.

RGB5_A1
The red, green, and blue channels are stored with 5 bits, and the alpha channel is stored with 1 bit.

RGBA
A texel contains red, green, blue, and alpha components.

RGBA4
The red, green, blue, and alpha channels are all stored with 4 bits.

STENCIL_INDEX8
The renderbuffer stores an 8-bit stencil component.

UNSIGNED_BYTE_TEXEL
Each texel is 4 * 8 bits.

UNSIGNED_SHORT_4_4_4_4
Each texel is 16 bits, with 4, 4-bit components.

UNSIGNED_SHORT_5_5_5_1
Each texel is 16 bits, with 5-bit/5-bit/5-bit/1-bit components.

UNSIGNED_SHORT_5_6_5
Each texel is 16 bits, with 5-bit/6-bit/5-bit components.

Traits

Image2d
A trait for types that that contain 2D image data, which can be uploaded to a texture using a Texture2dBuilder or using the gl.image_2d method.