Struct glitter::index_data::slice::Chunks

pub struct Chunks<'a, T> where T: 'a {
    // some fields omitted
}

An iterator over a slice in (non-overlapping) chunks (size elements at a time).

When the slice len is not evenly divided by the chunk size, the last slice of the iteration will be the remainder.