Struct glitter::index_data::slice::Iter

pub struct Iter<'a, T> where T: 'a {
    // some fields omitted
}

Immutable slice iterator

Methods

impl<'a, T> Iter<'a, T>

fn as_slice(&self) -> &'a [T]

View the underlying data as a subslice of the original data.

This has the same lifetime as the original slice, and so the iterator can continue to be used while this exists.