Trait glitter::index_data::IndexDatum

pub unsafe trait IndexDatum {
    fn index_datum_type() -> IndexDatumType;
}

A single value that can be treated as a single index.

Safety

This type will be transmuted to a slice according to the value returned by the index_datum_type method. Implementing this method incorrectly will lead to memory unsafety.

Required Methods

fn index_datum_type() -> IndexDatumType

Return the data type this datum corresponds to.

Safety

An instance of this type must match the size and memory layout specified by the returned IndexDatumType.