Module glitter::prelude

Re-exports essential extension traits. Everything exported in this module should be used anywhere that glitter is used.

Generally, this means that most modules that use glitter will start with the following:

use glitter;
use glitter::prelude::*;Run

Reexports

pub use context::{AContext, BufferContext, ArrayBufferContext, ElementArrayBufferContext, FramebufferContext, ContextFramebufferBuilderExt, ProgramContext, ContextProgramBuilderExt, RenderbufferContext, ContextRenderbufferBuilderExt, TextureBinding, ContextTextureBuilderExt, TextureUnit, TextureUnitBinding, ATextureUnitBinding, TextureUnitBinding2d, TextureUnitBindingCubeMap, TextureUnit0Context, TextureUnit1Context, TextureUnit2Context, TextureUnit3Context, TextureUnit4Context, TextureUnit5Context, TextureUnit6Context, TextureUnit7Context};

pub use context::ext::*;

pub use shader::ContextShaderBuilderExt;

pub use vertex_buffer::{VertexBufferContext, IndexBufferContext, ContextVertexBufferExt};

pub use types::GLObject;