Struct glitter::program::Program

pub struct Program {
    // some fields omitted
}

An OpenGL program object.

A program is used in OpenGL to customize the render pipeline. A program is composed of multiple Shader objects, each of which computes one step in the render pipeline, and feed data to the next program.

A program will automatically be deleted after going out of scope.

See also

gl.build_shader: Build and link a new program object

gl.create_program: Create a new, empty program object.

gl.use_program: Bind a program to use for drawing calls, returning a ProgramBinding type.