Struct glitter::program::ProgramUniform

pub struct ProgramUniform {
    pub gl_index: GLuint,
}

An OpenGL program uniform.

This type is used for getting and setting a program's uniform variables, which are constants used during a single draw call, but can be changed between draw calls.

See also

gl.get_uniform_location: Get a ProgramUnfirom from a uniform's name within a program.

gl.set_uniform: Set or change the value of a program's uniform.

Fields
gl_index: GLuint

The index of the the program uniform.