Struct glitter::renderbuffer::Renderbuffer

pub struct Renderbuffer { /* fields omitted */ }

An OpenGL renderbuffer object.

A renderbuffer is essentially a 2D image that's optimized to be a render target. Unlike a Texture2d, a renderbuffer is always created with uninitialized data. The only way to 'send' an image to a renderbuffer is to render to it by attaching it to a Framebuffer. A renderbuffer is most appropriate when rendering a final output image, where no extra postprocessing is required.

A renderbuffer will automatically be deleted after going out of scope.

See also

gl.build_renderbuffer: Build a new renderbuffer, using the RenderbufferBuilder type to setup storage options.

gl.gen_framebuffer: Create a new framebuffer with no attachment points.

gl.bind_renderbuffer: Bind a renderbuffer to a context, returning a RenderbufferBinding type