Module glitter::shader

Exposes the OpenGL Shader object and related types.

Structs

Shader
An OpenGL shader object.

ShaderBuilder
A safe interface for creating a shader with a source, and returning an error or panicking if there is an error.

Enums

ShaderType
The possible types of shader objects.

Constants

FRAGMENT_SHADER
A shader that is used for processing per-fragment (per-pixel) data.

VERTEX_SHADER
A shader that is used for processing per-vertex data.

Traits

ContextShaderBuilderExt
An extension trait for ContextOf that adds functions to build shaders using the ShaderBuilder interface.

ContextShaderExt
An extension trait that includes shader-related OpenGL methods.