Enum glitter::texture::TextureMipmapFilter

pub enum TextureMipmapFilter {
    Filter(TextureFilter),
    MipmapFilter {
        criterion: TextureFilter,
        mipmap: TextureFilter,
    },
}

Represents the different forms of texture filtering when using mipmaps.

Variants

Filter

Ignore mipmap values, and texture a pixel using a standard TextureFilter.

MipmapFilter

Select the two mipmaps that are closest to the size of pixel being filled, and sample each one according to criterion. Finally, the result will be computed either by taking the weighted average of each texel, or by selecting the value from the closer texel, according to mipmap.

Fields

criterion

The method to use to select the texels from a mipmap.

mipmap

The method to use to select the mipmaps.

Methods

impl TextureMipmapFilter

fn gl_enum(&self) -> GLenum

Convert a TextureMipmapFilter into a raw OpenGL enum value