Enum glitter::types::GLError

pub enum GLError {
    InvalidEnum,
    InvalidValue,
    InvalidOperation,
    InvalidFramebufferOperation,
    OutOfMemory,
    FramebufferError(GLFramebufferError),
    Message(String),
}

The various possible OpenGL errors.

Variants

InvalidEnum

Indicates that an unexpected enum value was passed to a function.

InvalidValue

Indicates that a function was passed an argument with a value that had an unexpected value (such as passing a negative value to a function that only expects positive arguments).

InvalidOperation

Indicates that a particular operation was attempted that is not allowed, often resulting from an unexpected pairing of arguments.

InvalidFramebufferOperation

Indicates that a particular operation attempted to use a framebuffer that is not framebuffer-complete.

OutOfMemory

Indicates that the OpenGL driver could not allocate enough memory to satisfy a request.

FramebufferError

Indicates a framebuffer-related error.

Message

Indicates an error with a message attached (such as a message from an info log, or an error message originating from glitter).