Module glitter::uniform_data

Contains types that represent uniform data, which is used for methods such as gl.set_uniform.

Enums

UniformDatumType
The basic types that can be used as uniform values in a program object.

UniformPrimitiveType
The basic value types that are composed in the UniformDatumTypes.

Traits

UniformData
A type that can be set to a uniform value in a program object, using the gl.set_uniform method. A UniformData type can be composed of one or more UniformDatums, and this is likely the type that user types will implement.

UniformDatum
A single uniform value, which corresponds to a single primitive GLSL uniform type.

UniformPrimitive
A single, basic value that can be composed to make a UniformDatum. Scalar values are an example of a UniformPrimitive.