Trait glitter::uniform_data::UniformDatum

pub unsafe trait UniformDatum {
    fn uniform_datum_type() -> UniformDatumType;
}

A single uniform value, which corresponds to a single primitive GLSL uniform type.

Safety

This type will be transmuted to a slice according to the value returned by the uniform_datum_type method. Implementing this method incorrectly will lead to memory unsafety.

Required Methods

fn uniform_datum_type() -> UniformDatumType

Return the data type this datum corresponds to.

Safety

An instance of this type must match the size and memory layout specified by the returned UniformDatumType.