Trait glitter::uniform_data::UniformPrimitive

pub unsafe trait UniformPrimitive {
    fn uniform_primitive_type() -> UniformPrimitiveType;
}

A single, basic value that can be composed to make a UniformDatum. Scalar values are an example of a UniformPrimitive.

Safety

This type will be transmuted to a slice according to the value returned by the uniform_primitive_type method. Implementing this method incorrectly will lead to memory unsafety.

Required Methods

fn uniform_primitive_type() -> UniformPrimitiveType

Return the data type this primitive corresponds to.

Safety

An instance of this type must be the size of the UniformPrimitiveType returned by this function.