Module glitter::vertex_buffer

Contains a higher-level abstraction for creating vertex and index buffer.

Structs

AttribBinder
A type used to perform operations on a list of program attributes. An instance of an AttribBinder can be created with the attrib_pointers! macro or with the new method.

AttribError
An error generated when trying to iterate over the lists of attributes in an AttribBinder.

IndexBuffer
A buffer that contains index data. In addition to storing a buffer, an IndexBuffer store a count of the amount of IndexData that has been buffered.

IndexBufferBinding
Represents an IndexBuffer that has been bound to a context.

VertexBuffer
A buffer that contains vertex data. In addition to storing a buffer, a VertexBuffer stores an AttribBinder and a count of the amount of VertexData that has been buffered.

VertexBufferBinding
Represents a VertexBuffer that has been bound to a context.

Enums

AttribAddError
An error generated when trying add an attribute to an AttribBinder using the AttribBinder::add method.

VertexBindError
An error generated when binding vertex attributes.

Traits

ContextVertexBufferExt
An extension trait that adds vertex buffer and index buffer-related methods to OpenGL contexts.

IndexBufferContext
An OpenGL context that can have an index buffer bound.

VertexBufferContext
An OpenGL context that can have a vertex buffer bound.