Trait glitter::vertex_buffer::VertexBufferContext

pub trait VertexBufferContext: ArrayBufferContext + Sized {
    fn bind_vertex_buffer<'a, V>(self, vbo: &'a mut VertexBuffer<V>) -> (VertexBufferBinding<V>, Self::Rest) where V: VertexData { ... }
}

An OpenGL context that can have a vertex buffer bound.

Note

Interanlly, vertex buffers are bound to the GL_ARRAY_BUFFER binding, so any context that has a free GL_ARRAY_BUFFER is a VertexBufferContext.

Provided Methods

fn bind_vertex_buffer<'a, V>(self, vbo: &'a mut VertexBuffer<V>) -> (VertexBufferBinding<V>, Self::Rest) where V: VertexData

Bind a vertex buffer to this context, returning a binding and a new context.