Struct glitter::context::texture_units::TextureUnitBindingOf

pub struct TextureUnitBindingOf<T2, TC> { /* fields omitted */ }

A texture unit that has been made active, and can have textures bound to it.

Methods

impl<T2, TC> TextureUnitBindingOf<T2, TC>

pub fn gl_idx(&self) -> u32

Get the index of the texture unit.

pub fn sampler(&self) -> TextureSampler

Get the current texture unit as a TextureSampler, which can be used to set a uniform variable.