Struct glitter::framebuffer::Framebuffer

pub struct Framebuffer {
    // some fields omitted
}

An OpenGL framebuffer object.

A framebuffer can best be thought of as a 'surface' that can be rendered to. A framebuffer will contain one or more attachment points, which point to the actual container where data will really be written when drawing. Generally, a framebuffer will use a Renderbuffer to draw a final image to the screen, and a Texture2d when some post-processing effects need to be applied before displaying the final image.

A framebuffer will automatically be deleted after going out of scope.

See also

gl.build_framebuffer: Build a new framebuffer, using the FramebufferBuilder type to set attachment points.

gl.gen_framebuffer: Create a new framebuffer with no attachment points.

gl.bind_framebuffer: Bind a framebuffer to a context, returning a FramebufferBinding type.