Trait glitter::index_data::IndexData

pub trait IndexData {
    fn index_datum_type() -> IndexDatumType;
    fn index_bytes(&self) -> &[u8];
    fn index_elements(&self) -> usize;
}

Indicates that a type can be coerced to a u8 slice that can be treated as a stream of index data.

Required Methods

fn index_datum_type() -> IndexDatumType

Return the data type that this index data should be treated as.

fn index_bytes(&self) -> &[u8]

Create a byte slice of index data from self.

fn index_elements(&self) -> usize

Return the number of index elements that self contains.